fix : 위치 권한 확인 alert 수정 (#144)
