+++ KCS/KCS/Presentation/Home/View/HomeViewController.swift
@@ -101,25 +101,6 @@ final class HomeViewController: UIViewController {
         return map
     }()
     
-    private let requestLocationServiceAlert: UIAlertController = {
-        let alertController = UIAlertController(
-            title: "위치 정보 이용",
-            message: "위치 서비스를 사용할 수 없습니다.\n디바이스의 '설정 > 개인정보 보호'에서 위치 서비스를 켜주세요.",
-            preferredStyle: .alert
-        )
-        let goSetting = UIAlertAction(title: "설정으로 이동", style: .destructive) { _ in
-            if let appSetting = URL(string: UIApplication.openSettingsURLString) {
-                UIApplication.shared.open(appSetting)
-            }
-        }
-        let cancel = UIAlertAction(title: "취소", style: .cancel)
-        
-        alertController.addAction(cancel)
-        alertController.addAction(goSetting)
-        
-        return alertController
-    }()
-    
     private lazy var refreshButton: RefreshButton = {
         let button = RefreshButton()
         button.translatesAutoresizingMaskIntoConstraints = false
@@ -380,7 +361,7 @@ private extension HomeViewController {
         viewModel.locationAuthorizationStatusDeniedOutput
             .bind { [weak self] _ in
                 guard let self = self else { return }
-                present(requestLocationServiceAlert, animated: true)
+                presentLocationAlert()
             }
             .disposed(by: disposeBag)
